fastapi/requirements.txt
fastapi>=0.95
uvicorn==0.21.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
SQLAlchemy==2.0.9
python-dotenv==1.0.0
